Phase 2.1: Refactor lit_model.py into modular components
Split the monolithic lit_model.py (1846 lines) into focused, maintainable modules:
- lit_model.py (539 lines, 71% reduction): Core Lightning orchestration
- deep_supervision.py (404 lines): Deep supervision and multi-task learning
- inference.py (868 lines): Sliding window inference, TTA, postprocessing
- debugging.py (173 lines): NaN detection and debugging utilities

Key improvements:
- Eliminated code duplication through delegation pattern
- Clear separation of concerns with single responsibility per module
- Improved maintainability and testability
- No backward compatibility methods (clean design per user requirement)

All modules pass Python syntax compilation. Integration tests pending.

Closes #2.1 in REFACTORING_PLAN.md
