+++ connectomics/lightning/debugging.py
@@ -1,0 +1,173 @@
+"""
+Debugging utilities for PyTorch Connectomics.
+
+This module implements:
+- NaN/Inf detection in activations
+- NaN/Inf detection in parameters and gradients
+- Forward hook management for intermediate layer inspection
+- Debug statistics collection and reporting
+"""
+
+from __future__ import annotations
+from typing import Dict, Any, Optional, Tuple
+
+import torch
+import torch.nn as nn
+
+from ..utils.debug_hooks import NaNDetectionHookManager
+
+
+class DebugManager:
+    """
+    Manager for debugging operations including NaN detection.
+
+    This class handles:
+    - Forward hooks for NaN/Inf detection in layer outputs
+    - Parameter and gradient inspection
+    - Statistics collection and reporting
+    - Interactive debugging support (pdb integration)
+
+    Args:
+        model: PyTorch model to debug (nn.Module)
+    """
+
+    def __init__(self, model: nn.Module):
+        self.model = model
+        self._hook_manager: Optional[NaNDetectionHookManager] = None
+
+    def enable_nan_hooks(
+        self,
+        debug_on_nan: bool = True,
+        verbose: bool = False,
+        layer_types: Optional[Tuple] = None,
+    ) -> NaNDetectionHookManager:
+        """
+        Enable forward hooks to detect NaN in intermediate layer outputs.
+
+        This attaches hooks to all layers in the model that will check for NaN/Inf
+        in layer outputs during forward pass. When NaN is detected, it will print
+        diagnostics and optionally enter the debugger.
+
+        Useful for debugging in pdb:
+            (Pdb) pl_module.enable_nan_hooks()
+            (Pdb) outputs = pl_module(batch['image'])
+            # Will stop at first layer producing NaN
+
+        Args:
+            debug_on_nan: If True, enter pdb when NaN detected (default: True)
+            verbose: If True, print stats for every layer (slow, default: False)
+            layer_types: Tuple of layer types to hook (default: all common layers)
+
+        Returns:
+            NaNDetectionHookManager instance
+        """
+        if self._hook_manager is not None:
+            print("⚠️  Hooks already enabled. Call disable_nan_hooks() first.")
+            return self._hook_manager
+
+        self._hook_manager = NaNDetectionHookManager(
+            model=self.model,
+            debug_on_nan=debug_on_nan,
+            verbose=verbose,
+            collect_stats=True,
+            layer_types=layer_types,
+        )
+
+        return self._hook_manager
+
+    def disable_nan_hooks(self):
+        """
+        Disable forward hooks for NaN detection.
+
+        Removes all hooks that were attached by enable_nan_hooks().
+        """
+        if self._hook_manager is not None:
+            self._hook_manager.remove_hooks()
+            self._hook_manager = None
+        else:
+            print("⚠️  No hooks to remove.")
+
+    def get_hook_stats(self) -> Optional[Dict[str, Dict[str, Any]]]:
+        """
+        Get statistics from NaN detection hooks.
+
+        Returns:
+            Dictionary mapping layer names to their statistics, or None if hooks not enabled
+        """
+        if self._hook_manager is not None:
+            return self._hook_manager.get_stats()
+        else:
+            print("⚠️  Hooks not enabled. Call enable_nan_hooks() first.")
+            return None
+
+    def print_hook_summary(self):
+        """
+        Print summary of NaN detection hook statistics.
+
+        Shows which layers detected NaN/Inf and how many times.
+        """
+        if self._hook_manager is not None:
+            self._hook_manager.print_summary()
+        else:
+            print("⚠️  Hooks not enabled. Call enable_nan_hooks() first.")
+
+    def check_for_nan(self, check_grads: bool = True, verbose: bool = True) -> dict:
+        """
+        Debug utility to check for NaN/Inf in model parameters and gradients.
+
+        Useful when debugging in pdb. Call as: pl_module.check_for_nan()
+
+        Args:
+            check_grads: Also check gradients
+            verbose: Print detailed information
+
+        Returns:
+            Dictionary with NaN/Inf information
+        """
+        nan_params = []
+        inf_params = []
+        nan_grads = []
+        inf_grads = []
+
+        for name, param in self.model.named_parameters():
+            # Check parameters
+            if torch.isnan(param).any():
+                nan_params.append((name, param.shape))
+                if verbose:
+                    print(f"⚠️  NaN in parameter: {name}, shape={param.shape}")
+            if torch.isinf(param).any():
+                inf_params.append((name, param.shape))
+                if verbose:
+                    print(f"⚠️  Inf in parameter: {name}, shape={param.shape}")
+
+            # Check gradients
+            if check_grads and param.grad is not None:
+                if torch.isnan(param.grad).any():
+                    nan_grads.append((name, param.grad.shape))
+                    if verbose:
+                        print(f"⚠️  NaN in gradient: {name}, shape={param.grad.shape}")
+                if torch.isinf(param.grad).any():
+                    inf_grads.append((name, param.grad.shape))
+                    if verbose:
+                        print(f"⚠️  Inf in gradient: {name}, shape={param.grad.shape}")
+
+        result = {
+            'nan_params': nan_params,
+            'inf_params': inf_params,
+            'nan_grads': nan_grads,
+            'inf_grads': inf_grads,
+            'has_nan': len(nan_params) > 0 or len(nan_grads) > 0,
+            'has_inf': len(inf_params) > 0 or len(inf_grads) > 0,
+        }
+
+        if verbose:
+            if not result['has_nan'] and not result['has_inf']:
+                print("✅ No NaN/Inf found in parameters or gradients")
+            else:
+                print(f"\n📊 Summary:")
+                print(f"   NaN parameters: {len(nan_params)}")
+                print(f"   Inf parameters: {len(inf_params)}")
+                print(f"   NaN gradients: {len(nan_grads)}")
+                print(f"   Inf gradients: {len(inf_grads)}")
+
+        return result
